+++ frontend/src/features/manageVehicles/components/list/List.scss
@@ -2,16 +2,3 @@
     margin-top: 10px;
     overflow: hidden;
 }
-
-.options-container {
-    display: flex;
-    padding: 7px 0 7px 0;
-}
-
-.csv-container {
-    margin-left: auto;
-}
-
-.trash {
-    padding: 0px 30px;
-}
