+++ frontend/src/features/manageVehicles/components/options/CSVOptions.tsx
@@ -3,7 +3,7 @@ import IconButton from '@mui/material/IconButton';
 import Menu from '@mui/material/Menu';
 import MenuItem from '@mui/material/MenuItem';
 import MoreVertIcon from '@mui/icons-material/MoreVert';
-import "./List.scss";
+import "./Options.scss";
 
 const options = [
   'Import CSV',
