+++ frontend/src/features/manageVehicles/components/options/Options.tsx
@@ -1,10 +1,10 @@
-import "./List.scss";
+import "./Options.scss";
 import { CSVOptions } from "./CSVOptions";
 import { Search } from "./Search";
 import { Filter } from "./Filter";
 import { Trash } from "./Trash";
 
-export const OptionsBar = () => {
+export const Options = () => {
   return (
     <div className="options-container">
       <Search />
