+++ frontend/src/features/manageVehicles/components/options/Trash.tsx
